requirements.txt
django==4.1.7
django-rest-framework==0.1.0
django-environ==0.10.0
django-celery-results==2.4.0
celery==5.2.7
redis==4.5.1
psycopg2-binary==2.9.5
loguru==0.6.0
gunicorn==20.1.0
